Moved a module to utilities package
