+++ nmt_chainer/__main__.py
@@ -6,7 +6,7 @@ import nmt_chainer.translation.eval_config as eval_module
 import nmt_chainer.dataprocessing.make_data as make_data
 import nmt_chainer.translation.server_arg_parsing as server
 
-import nmt_chainer.utils_command as utils_command
+import nmt_chainer.utilities.utils_command as utils_command
 
 from nmt_chainer.utilities import versioning_tools
 import sys
@@ -53,27 +53,22 @@ def main(arguments = None):
     # create the parser for the "make_data" command
     parser_make_data = subparsers.add_parser('make_data', description= "Prepare data for training.", help = "Prepare data for training", formatter_class=argparse.ArgumentDefaultsHelpFormatter)
     make_data.define_parser(parser_make_data)
-#     parser_make_data.set_defaults(func_str = "make_data")
     
     # create the parser for the "train" command
     parser_train = subparsers.add_parser('train', description= "Train a model.", help = "Train a model", formatter_class=argparse.ArgumentDefaultsHelpFormatter)
     train.define_parser(parser_train)
-#     parser_train.set_defaults(func_str = "train")
     
     # create the parser for the "eval" command
     parser_eval = subparsers.add_parser('eval', description= "Use a model.", help = "Use a model", formatter_class=argparse.ArgumentDefaultsHelpFormatter)
     eval_module.define_parser(parser_eval)
-#     parser_eval.set_defaults(func_str = "eval")
     
     # create the parser for the "server" command
     parser_server = subparsers.add_parser('server', description= "Launch a server.", help = "Launch a server",
                                           formatter_class=argparse.ArgumentDefaultsHelpFormatter)
     server.define_parser(parser_server)
-#     parser_server.set_defaults(func_str = "server")
 
     # create the parser for the "version" command
     parser_version = subparsers.add_parser('version', description= "Get version infos.", help = "Get version infos", formatter_class=argparse.ArgumentDefaultsHelpFormatter)
-#     parser_version.set_defaults(func_str = "version")
     
     parser_utils = subparsers.add_parser('utils', description= "Call a utility script.", help = "Call a utility script", formatter_class=argparse.ArgumentDefaultsHelpFormatter)
     utils_command.define_parser(parser_utils)
