+++ nmt_chainer/utilities/utils_command.py
